Don't try to use filename as libvirt VM name
When creating a VM from an ISO that is unknown from libosinfo, we
use the ISO filename as the libvirt VM name. However, this has
a number of issues. First, libvirt documentation says that "The
content of the name element provides a short name for the virtual
machine.  This name should consist only of alpha-numeric characters"
The ISO filename could contain any characters, so is not a really
good fit for a VM name. Moreover, libvirt then uses this name
to create some temporary filenames with a size limit, so if the
ISO name gets too long, starting the VM will not be possible because
this filename is too long:

(gnome-boxes:6917): Boxes-DEBUG: app.vala:812: connect display failed:
Unable to start domain: internal error Monitor path
/home/teuf/.config/libvirt/qemu/lib/rhel-server-6.3-x86_64-dvd.iso?__gda__=xxxxxxxxxx_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx&ext=.iso.monitor
too big for destination
This commit uses 'boxes-unknown' as the base name for the VM instead
of the filename to avoid these issues. From an UI point of view, this
should not change anything as we show the VM title in the UI, not its
name.

https://bugzilla.gnome.org/show_bug.cgi?id=689211