installer: Use libosinfo APIs for the automated installations
Let's drop all OSes-specific automation code in favor of new
automated installation API provided by libosinfo.

Known issues:

1. Direct boot commandline should come from libosinfo.
2. We still have os-specific code for deciding what disk to use for
   unatteded files.
3. This breaks viostor drivers installation.

I'll work on #1 next but hardcoded value for now isn't that bad as what
we do currently isn't any different. Same for #2 and IMO the solution for
that would go under (yet to be created) libvirt-builder.

As for #3, a following patch in this series fixes that using new
libosinfo API.

co-author: Zeeshan Ali (Khattak) <zeeshanak@gnome.org>

https://bugzilla.gnome.org/show_bug.cgi?id=676537