Don't try to use unsupported languages in autoinstalls
Now that libosinfo has an API to report which languages are
supported by an ISO, we can fix a long-standing bug with Windows
unattended installations: Windows ISOs only support one language
most of the time, and we need to specify this language as the
one to use for installation, otherwise Windows will ask a question.

This commit will first check if the current installer supports one
of the languages from the user locale and use that if found.
Otherwise it will fallback to using one of the languages supported
by the media. If the media does not have any associated language,
it will use the user locale as the installation language.

https://bugzilla.gnome.org/show_bug.cgi?id=685826