installer: Prefix driver files with location's MD5
Instead of OS ID and architecture, lets just prefix driver files with
MD5 sum of their location. This way we not only guarantee uniqueness but
also ensure that our cache is up2date. The latter assumes that location
is versioned (as it should be) so thats something we need to keep in
mind when updating drivers in libosinfo DB.

https://bugzilla.gnome.org/show_bug.cgi?id=690169