installer: Use bridge net if system libvirt is available
Use bridge networking in new VMs, if available. This network is much
faster than the user networking.

Unfortunately bridge networking being available implies running system
libvirtd. However since system libvirtd is typically started on boot when
libvirt is installed, we can perhaps rely on it being available at
subsequent launch of the created VM in future?

NB: We don't need to actually connect to system libvirt for this to work.

https://bugzilla.gnome.org/show_bug.cgi?id=677688