Remove deprecated clutter stuff from Revealer
This switches from using direct "animate" calls to using implicit
animations. Also, we animate the translation rather than the position
of the items as this is unrelated to the allocation and thus avoids
lots of resize calculations.

We also use a new custom layout manager instead of the bin as binlayout
allocates a smaller child when we change the height (if resize==true)
rather than clipping it.

https://bugzilla.gnome.org/show_bug.cgi?id=691078