downloader: (Try to) Support non-HTTP cases
This is not guaranteed to work for every kind of filesystem but at least
now we don't assume HTTP blindly and support caching from local
filesystem. Its a perfectly valid usecase for distros to provide product
logos or driver files from local filesystem and with this patch we'll
support that.

https://bugzilla.gnome.org/show_bug.cgi?id=691823