Don't allow smaller window than the UI requires
We activate the GtkClutter.Embed use-layout-size mode to ensure this.

With this it may happen that the window size grows when you switch
to a new mode, like the wizard. However, since the UI packing is
pretty good these days this really only happens for really tiny
window sizes which are not very common. And any way, its far better
than just cutting off the UI.

https://bugzilla.gnome.org/show_bug.cgi?id=678960