Proper support for command line args in secondary process
We now handle most of the command line parsing (except a few
specialized things) in the primary instance, by implementing the
command_line signal of GApplication. This means we now support
arguments like --open-uuid in the non-primary instance, as needed
for e.g. gnome-shell search results.

This additionally moves the UI initialization and window to
the activate state so that it only happens in the primary instance,
and it doesn't happen if we exit early in the command_line handler.

Unfortunately we can't use the automatic --help support in GOptions
anymore as that prints to/exists the primary instance rather than
the calling one, so we have to now manually do that.

https://bugzilla.gnome.org/show_bug.cgi?id=692306