Be more robust when restarting installations
When Boxes is restarted, it will automatically restart domains that
were doing an unattended installation. While doing that, the
code assumes the ISO image used to start the installation will still
be present in the domain XML configuration. However, this is not
necessarily true as it's possible that the installation was completed
outside of Boxes (using virt-viewer for example), and that the
user then ejected the ISO image using virsh or virt-manager.
In such a situation, Boxes will crash at startup, which should
never happen.
This commit adds null checks here and there and fixes the issue, but
hopefully a more elegant solution can be found...

https://bugzilla.gnome.org/show_bug.cgi?id=693134