installer-media: Fetch original volume ID from udev
udev replaces spaces (' ') with underscores ('_') in 'ID_FS_LABEL'
property. Since there is no way to know if an underscore was
originally a space or not in the source volume ID, we better use
'ID_FS_LABEL_ENC' instead. In that property spaces are escaped by '\x20'
and hence we can be much more certain about it being originally a space.

https://bugzilla.gnome.org/show_bug.cgi?id=694141