installer: Remove redundant loop
Once the drivers are setup, they are setup for all scripts that can use
them. There is no need to keep setting them up for every script.

Also if there is going to be multiple scripts in the desktop profile
supporting same type (pre- or post-installable) of drivers, Boxes will be
copying them multiple times. That is not a big issue in case of
pre-installable drivers but in case of post-installable drivers, the
mkisofs commandline we use to setup postinstallation drivers will fail.

This patch fixes that potential issue as well.

https://bugzilla.gnome.org/show_bug.cgi?id=694872