Don't pass GError::message directly to got_error
The error message stored in a GError generally comes from underlying
libraries, and is not good enough to be shown to the user most of
the time. This commit replaces uses of got_error with GError::message
with more generic messages complemented by a debug() with the
actual error message.

I noticed this after getting a
"Error connecting XXX: no free USB channel" in the UI when trying to
redirect an USB device.

https://bugzilla.gnome.org/show_bug.cgi?id=695111