Hide USB devices from properties when usbredir is disabled
We currently show USB devices to be redirected in the 'devices'
tab of boxes properties even when Boxes was compiled without
--enable-usbredir. When trying to use such a device on a VM without
USB redirection support, the redirection would just fail with an
error message with no way for the user to fix it.
Showing the devices in this case is unintended, so let's hide them.

https://bugzilla.gnome.org/show_bug.cgi?id=695114