Dynamically refresh USB devices properties
Currently, when an USB device is plugged/unplugged, if the user
is looking at the devices property page of a box, the list of devices
does not automatically add/remove the new device.

The IPropertiesProvider interface does not provide an easy way to
dynamically add/remove properties, however given one property,
it's possible to trigger a refresh of the whole property page by
emitting the Property::refresh-properties signal.
This commit uses the 'automatic USB redirection' property to trigger
a page refresh as it will always be present when USB redirection is
supported.

https://bugzilla.gnome.org/show_bug.cgi?id=695207