Order the list of USB devices
When USB devices are unplugged and replugged, they won't
always reappear in the same place, which can be disturbing. This
commit alphabetically sort that list so that a given device always
reappear at the same place. The sort could be made more sophisticated
in the future if we want to group the devices by type, this kind of
things.

https://bugzilla.gnome.org/show_bug.cgi?id=695207