Use ref for IPropertyProvider::get_properties 'flag' arg
It seems this is what LibvirtMachineProperties::get_properties()
expects as it is modifying its 'flag' argument to indicate USB support
is missing.
SpiceDisplay::get_properties() is then checking for the presence
of this flag, and hides the list of USB devices if the box
does not have USB redirection support.

However, without using 'ref' on the get_properties 'flag' argument,
the change made in LibvirtMachineProperties won't get propagated
to the caller (this is LibvirtMachine::get_properties), and then
SpiceDisplay::get_properties will get unmodified flags and won't
know it should disable USB support.

https://bugzilla.gnome.org/show_bug.cgi?id=695207