installer: Don't automatically download drivers for all media
With this patch, we no longer start to download drivers as part of
UnattendedInstaller construction but rather do it as part of a new
general InstallerMedia.prepare method that is called as part of wizard's
'preparation' stage.

The reason for this change is that us automatically downloading all
drivers for all ISOs/medias users have, could easily congest a slow
network. So now we only download drivers for a media/OS when users
create their first VM against the media/OS in question.

This also fixes the issue of population of wizard's media menu getting
blocked by driver downloads.

https://bugzilla.gnome.org/show_bug.cgi?id=695335