Disallow changing of system props during install
Changing of system properties require a reboot and rebooting the machine
in the middle of an installation/live session is not a good idea. More so
since we track end of installation/live session through number of
reboots, especially in case of live session where we automatically remove
the machine on shutdown/reboot if no installation was performed.

https://bugzilla.gnome.org/show_bug.cgi?id=690779