installer: Disable driver signing for unsigned drivers
In the next release (0.2.6) of libosinfo, driver signing checks in
scripts will have to be explicitly disabled if we want to use unsigned
drivers. We actually shouldn't be doing this since this means disabling
a security feature of Windows *permanently* but unfortunately QXL drivers
for windows are not signed and they are essential for an impressive user
experience.

To make things slightly better, we only do this if there is an unsigned
driver to be installed.

With this change QXL driver will now actually be functional in Windows XP
and Windows 7.

https://bugzilla.gnome.org/show_bug.cgi?id=691546