Use configure.ac's VERSION instead of BUILD_VERSION
Boxes 3.8.0 reports 3.7.91 as its version even though .tarball-version
has the correct 3.8.0 version. However .version contains 3.7.91,
and this is what is currently used to generate BUILD_VERSION, which
Boxes then uses to report its version in the UI/in --version.
.
Since config.h already provides us with a VERSION #define, it's
better to use this one rather than guessing a version on our
own.

https://bugzilla.gnome.org/show_bug.cgi?id=696709