app: Avoid adding duplicate sources
Without this fix, we were adding the default 'QEMU Session' source
twice. It hasn't been a real problem so far but it will be after we
starting deleting stale box configs in the following patches.

We are now simply moving the add_collection_source out of
foreach_filename_from_dir's callback as we can't yield in there because
of bug#604827 and removing the explicit addition of 'QEMU Session'
source as it becomes redundant with this patch.

We were launching add_collection_source *async* for each source we found.
Since add_collection_source() checks if source is already added *before*
defering to mainloop, we could have a scenario like this:

1. source 'A' is found and add_collection source() defers to mainloop.
2. source 'B' (which is a duplicate of 'A') is found and
add_collection_source is launched for it too.
3. The add_collection source() from #1 has not yet added source 'A' so
the check for duplicate will fail and we'll still have a duplicated
source added.

https://bugzilla.gnome.org/show_bug.cgi?id=683489