wizard: Focus sensitive Continue/Create buttons
Continue/Create button should have the focus when they are sensitive.
Them being sensitive means that everything is ready for us to proceed to
next step and so we should make it very easy for user to quickly move to
next step. In most cases, this will enable users to create VMs very
quickly IMO.

https://bugzilla.gnome.org/show_bug.cgi?id=664218