Make sure wizard buttons use underlined chars
Because of libgd bgo#698289, creating a button with a label containing a
"_" and setting use_underline to TRUE afterwards don't make
GdHeadersimpleButton reinterpret the label to remove the '_' character
and use it as an underline mark.

Setting the label to the value it already has right after setting
use_underline works around this issue.

https://bugzilla.gnome.org/show_bug.cgi?id=698287