display-page: Show overlay toolbar until mouse moves
Based on the user feedback, it seems that its very easy to miss the fact
that there is a toolbar available in fullscreen mode, especially now
that we go to fullscreen automatically. Better keep the overlay toolbar
visible until after 1 second of user moving the mouse to ensure that they
at least know it exists and where.

https://bugzilla.gnome.org/show_bug.cgi?id=700574