Make sure newly created VMs are marked as used recently
When creating a VM for an unattended install, we don't connect
to its display automatically but run it minimized. This means
it never gets marked as recently used automatically as this is
done in the Display class. This commit sets Machine::access_last_time
when starting a VM through VMCreator::launch_vm() to ensure
the newly created VM appears first in the collection view even
if it's an express install.

https://bugzilla.gnome.org/show_bug.cgi?id=700777