installer: Introducing clean_up_preparation_cache method
Add a separate method for cleaning up any cache created by 'prepare'
method that need to be kept around until VM is created and launched. This
is kept separate from 'clean_up' method as that is meant to clean data
that is needed until VM is completely installed. To ensure
clean_up_preparation_cache is always called, we also call it from
'clean_up' method.

This fixes the issue of setup->review->setup->review wizard transition,
while keeping express installation selected, resulting in an error that
was triggered by us not passing any file arguments to mkisofs command as
we were already emptying the list of files.

https://bugzilla.gnome.org/show_bug.cgi?id=700470