Make sure new remote VMs are marked as used recently
When creating a VM for a remote connection (SPICE/VNC) we don't
connect to its display automatically but run it minimized. This means
it never gets marked as recently used automatically as this is
done in the Display class. This commit sets Machine::access_last_time
when creating a RemoteMachine instance to ensure the newly created VM
appears first in the collection view.

https://bugzilla.gnome.org/show_bug.cgi?id=700777