ovirt: Fix cancelling of authentication
When authentication to an oVirt broker is cancelled,
yield OvirtBroker::add_source() would never return until the
authentication times out. Together with the changes
6b95a50c2 'app: Avoid adding duplicate sources', this could cause
the UI to appear empty for the duration of the timeout if an oVirt
source is created first, and if the user cancelled the connection
(or even if the connection takes quite some time).

https://bugzilla.gnome.org/show_bug.cgi?id=683489