Asynchronously load VM sources
The fix for bug #683489 changed the various VM sources to be loaded
sequentially. As some sources can be slower than others (remote oVirt
VS local libvirt connection), this means a slow source could block
processing of a faster source, causing the UI to be empty for longer
than it should.

This commit follows one of Zeeshan's suggestion to make that processing
asynchronous again.