libvirt-machine-props,wizard: Hide storage volume when importing
When machine is in the process of being imported, hide its storage
volume from UI because the actual storage capacity will only be
known after import is finished.

This actually make me wonder if we should import the media during
wizard's 'preparation' step instead, keeping in mind that that we intend
to do downloading of medias in there as well.

https://bugzilla.gnome.org/show_bug.cgi?id=690757