Fix glib critical at the end of non-express installs
At the end of installations, regardless of if they are express
installs or not, setup_post_install_domain_config is called.
It unconditionally tries to remove 'disk_file' from the VM configuration.
However, this member is only set for express installs, it's NULL
for non-express installs, so we get glib criticals in the console
when trying to use it as if it was non-NULL.
This commit avoids this issue by checking if it's NULL before using
it, as is done in most of unattended-installer.vala

https://bugzilla.gnome.org/show_bug.cgi?id=701695