Handle SPICE connection errors
When an error occurs during SPICE connection, a channel-event event
is emitted, and its type is SPICE_CHANNEL_ERROR_XXX. Boxes currently
catches some channel-event events, but does not do anything in case
of errors. When such an error occurs (for example because of a wrong
remote host certificate), Boxes UI is stuck on a centered half zoomed
screen, and does not do anything.
By emitting the disconnected() signal when such an error is caught, we
ensure we will go back to the collection view.

https://bugzilla.gnome.org/show_bug.cgi?id=701753