selectionbar: Don't let user delete under-construction VMs
While the ideal solution is probably to make it possible for user to
delete under construciton VMs too, it would be pretty difficult to
implement with us launching external processes as part of such
construction. Keeping in mind also that this only affects imported (from
installed media) VMs and importing usually takes much less time than
installations, I think this is not bad at least as a temporary work
around.

https://bugzilla.gnome.org/show_bug.cgi?id=690757