collection-view: Indicate under-construction boxes
Use the new libgd API to indicate under-construction boxes by showing a
spinner on top of the box thumbnail.

Later we'll use the same API to show progressbar for import case.
Unfortunately we can't do the same for installation case as we don't have
any reliable means to determine progess of installations.