wizard: Don't hang on failure to handle a URL
Boxes would just hang in wizard's preparation stage if you pass a path
to a file on a locally mounted remote share (e.g SMB). Although the real
solution is to be able to handle those, we should have this in place
just in case this codepath is walked for some other reason.

https://bugzilla.gnome.org/show_bug.cgi?id=688798