selectionbar: Fix selection bug on favorites
Selecting a machine not in favorites, while there is a machine that is
in favorites and selected, adds all selected items to favorites. This
happens because the active property is changed during the check for
selection, emitting a "clicked" signal.

This patch will block the "clicked" handler for favorite button while
changing this property.

https://bugzilla.gnome.org/show_bug.cgi?id=707076