display-page: Gd.MainToolbar -> Gtk.HeaderBar
Two reqression from this one:

* Dragging of titlebar to unfulscreen doesn't work even though i do:

-        add_events (Gdk.EventMask.POINTER_MOTION_MASK);
+        add_events (Gdk.EventMask.POINTER_MOTION_MASK |
+                    Gdk.EventMask.BUTTON_PRESS_MASK |
+                    Gdk.EventMask.BUTTON_RELEASE_MASK);
+

The header bar just won't receive any events. :( I also noticed that
this has been only working for OSD toolbar we show when server-side mouse
is available but not for the non-OSD toolbar. Keeping this fact in mind
and that there still is the unfullscreen button in an obvious place, I
don't think this is a very major regression. Having said that, I'll
hopefully fix it soon.

* The restore button in OSD toolbar looks chopped on the left side.
  Should be a matter of tuning the theme.

https://bugzilla.gnome.org/show_bug.cgi?id=706414