wizard: Put a box around 'no kvm' infobar
For some weird unknown reason, toggling the visibility of the infobar
itself doesn't exactly work. The bar remains visible even after setting
visibility to 'false' but its only visible as a line. Likely some bug in
clutter-gtk but I failed to reproduce it outside Boxes. :(

So we put the infobar in a box and toggle the visibility of that
instead.