installed-media: Special gnome-continuous images support
This patch links gnome-continuous images to appr. libosinfo's OS entry
so that Boxes gives it a nice name and assign supported devices. This
also fixes the issue of nonfunctional networking in VM created for these
images.

Ideally we shouldn't need to do this ourselves and libosinfo should do
the detection/mapping or we should look into libguestfs for this but
that would be a lot of work and it will not be very useful when we add
support for OVF files and gnome-continuous start to produce those. So
for now we go with this hack, which is not a that bad.

https://bugzilla.gnome.org/show_bug.cgi?id=706012