iso-extractor: Don't use fuseiso
Instead of using unmaintained fuseiso for extracting files from ISOs,
use iso-read command instead. This also saves us from having to mount
and unmount the ISO.

This implies that we now depend on libcdio[1]. The library itself is
GPLv3 so we can't link to it from our LGPLv2 app so we launch an external
utility provided (iso-read) by this library to avoid licensing issues.

[1] http://www.gnu.org/software/libcdio/