Revert "libvirt-machine-props: Allow changing network iface"
This reverts commit b2128c245b0d58fb8cf22f21f8a0861aba2b0fa5.

Turns out that bridge network setup through 'virbr0' interface actually
means a NAT setup where guests are on a separate private (internal to
host) network. This means that such a setup is as safe as 'user'
networking and therefore providing the option to change between the two
does not make much sense.

It might make sense to allow settting up a proper bridge network against
a particular interface in future though.

https://bugzilla.gnome.org/show_bug.cgi?id=720399