properties: Derive from Gtk.Notebook
Now that we are going more from containership to inheritance in other
widget classes to be able to make use of UI templates, lets do the same
for Properties class for consistency.

Moving this class' UI setup to .ui won't be so easy as most of the
contents are dynamically created.