machine: Don't use clutter actor
The animations of flying the machine display/screenshot around through
the whole UI is the main use of clutter/clutter-gtk for Boxes. Once this
usage is remove, we can easily remove rest of the usage without having
to break any UI freezes.

This implies that we remove the animations all together. We'll shortly
introduce animation for switch of view at least, which is possible using
gtk-alone.

Known issue: This somehow breaks connecting to a displays requiring
password. I spent many hours trying to find out why but the only thing I
found was that Spice.Display.ready property never becomes 'true' in this
case. I have also tried hardcoding the password to check if this is about
some lingering state from the first attempt to connect to display
without password but that did not help either.