app: Don't make view invisible on launching a box
When launching a box because of commandline argument, we were making the
view invisible. We used to make the actor holding the view invisible
instead and that made sense but now that view is not inside an actor, we
only switch the stack pages intead. So the removed code doesn't just not
make sense anymore but its also that it doesnt' work and only amounts to
Boxes show a spinner forevever if a commandline argument is passed.