desktop: Don't register for 'x-content/bootable-media'
Registering for this mimetype currently means that user is told that
Boxes can handle any device with that mimetype and given an easy to
launch Boxes for such devices. However, Boxes can:

1. not handle CDROM devics at all because of Qemu limitations.
2. only handle devices if user has read permissions on it and normally
   user does not for security reasons.

This means it will often happen so that user choose to launch Boxes,
after given an notification from shell/nautilus, and then end-up in an
error from Boxes or even worse, from the guest OS.

So we better not advertise Boxes as handler for this mimetype
until/unless we have a solution (working out of the box) for these
issues.