wizard-source: Only make target entries visible
To make media entries and system libvirt import button visible, we were
making their parent main_vbox and all its chilren visible. Thats not the
right thing to do and I believe was only a workaround for some weird
clutter-gtk or gtk+ bug that doesn't seem to be needed anymore.

This fixes the issue of 'Import 0 boxes from system broker' button
showing up in the wizard.

https://bugzilla.gnome.org/show_bug.cgi?id=726970