unattended-installer: Use correct cache path
d4bd651 caused a regression, where unattended installation setup failed
for OSes for which we download drivers (win7 and winxp) if the driver
files only exist in system cache since the code still tried to use the
user cache for all situations.

This patches fixes the issue by using the cache where there the file
exists/is downloaded.