vm-creator: Ensure VM name is <= 15 characters long
Commit 4dd0b75 put variant id into the VM names (and therefore
hostnames) to make it more likely to assign a meaningful name to VMs but
it also broke express installation for OSs with variants name that are
long and have very strict restrictions on number of characters in
hostnames, e.g win7 professional.

This patch fixes express installation of such OSes by ensuring that VM
name is not longer than 15 characters.