spice-display: PROPERTIES shouldn't unset resize-guest
Don't unset 'resize-guest' property of Spice.Display when going to
properties view as its not just redundant but also makes the display
grow or shrink to fit guest resolution, which could easily go beyond the
screen even.

We don't want our window to grow or shrink on its own and especially to
grow beyond the screen. If user was able to set a particular size on the
window, Boxes must not change it on its own as that would be contrary to
what user expects.

This issue didn't realize before commit 1854c96, where we made the
GtkStack holding most of the views, homogenous.

https://bugzilla.gnome.org/show_bug.cgi?id=730158