app-window: Disconnect machine on collection page
Moves the cancellation of the current machine's conection when the UI is
in collection state from App to AppWindow. As the 'current-item' property
is moving from App to AppWindow and as the UI state will too, the
cancellation should move. Also moves the setting of the current item's
state from App to AppWindow.

This is needed to drop the use of AppWindow singleton and therefore to
make multiple windows possible.

https://bugzilla.gnome.org/show_bug.cgi?id=732098