app: Remove 'current-item' prop
Removes the 'current-item' property from App. Also makes the
'current-item' property of AppWindow an attribute as it was only needed
to bind current items during the move. It finishes its move from App to
AppWindow.

This is needed to drop the use of AppWindow singleton and therefore to
make multiple windows possible.

https://bugzilla.gnome.org/show_bug.cgi?id=732098