collection-view: Setup UI once window is constructed
This is part of the move of the 'selection-mode' property from App to
AppWindow. This adds setup_ui() to the CollectionView class to ensure that
the connection to the notification of 'selection-mode' is done once the
window is constructed.

This is needed to drop the use of AppWindow singleton and therefore to
make multiple windows possible.

https://bugzilla.gnome.org/show_bug.cgi?id=732098