collection-toolbar: Move 'ui-state' connection to setup_ui()
This moves the connection to the App's 'ui-state' notifications from the
constuctor to setup_ui (). UIState will no more be implemented by App but
rather by AppWindow, so we must ensure that the window is constructed when
trying to connect to 'ui-state' notifications.

This is needed to drop the use of AppWindow singleton and therefore to
make multiple windows possible.

https://bugzilla.gnome.org/show_bug.cgi?id=732098