machine: Don't switch away from properties on display disconnection
If the user reverts to a snapshots and the domain is running, it will
temporarily be paused, causing its display to disconnect. We normally
switch to the collections view, but let's not do it if the user is
currently in the properties view since the state change is just normal
when reverting to a snapshot.

https://bugzilla.gnome.org/show_bug.cgi?id=710306