libvirt-machine-props: Fix max value of max disk
As Boxes only allocates the used space of a volume (instead of its
capacity), the capacity itself should not be taken into account when
calculating the max disk space and is better to use the volume's
allocation instead:
    max disk space = volume's allocation + available space in the pool

However it introduces an issue where the minimum storage value can be
bigger or equal to the maximum storage value. In this case lets just
show the current capacity as a label and not allow to change the value.

https://bugzilla.gnome.org/show_bug.cgi?id=726066