remote-machine: Make use of bidirectional prop binding
We were (manually) setting the name of the underlying source on change
of machine name while having a unidirectional binding from source's name
to machine name. This creates an endless loop and hangs Boxes when user
modifies a remote machine's name.

Lets just create a bidirectional binding and let GObject take care of
keeping name of both objects in sync.