libvirt-machine: Reset window on display reconnect
When reconnecting the display, we first disconnect the display and then
connect it again. Disconnecting however sets the window of a Machine to
null which we want to avoid in the case of reconnecting. Since Machine
defaults to the application's main_window if the machine's window is
unset, reconnecting the display caused the machine to be displayed in
the main_window and not in the window the user opened specifically for
it.

We fix this by saving the window of the machine, then reconnecting the
display and then just reassigning the window to the machine again.

https://bugzilla.gnome.org/show_bug.cgi?id=736081