downloader: Download to temporary location first
Instead of directly downloading to the cached path, download files first
to a temporary location. The filename has a '~' at the end so its
obvious that its a not regular file.

This fixes the issue of Boxes trying to use partially downloaded
files. Currently Boxes doesn't ensure that partially downloaded file is
deleted before exitting but even if it did, there will always been
crashes.

https://bugzilla.gnome.org/show_bug.cgi?id=738016