props-toolbar: Only return to collection if prev state was display
In commit commit b68642b we started to always return to 'collection' UI
state on user clicking back button if VM is not running. This was to fix
an issue that is specific to display->properties->display transition and
it breaks for wizard->properties->wizard transition. Lets fix the latter
by making sure we only return to collection state if previous UI state
was 'display'.

https://bugzilla.gnome.org/show_bug.cgi?id=738851