snapshot-list-row: Start machine on revert if prev state was DISPLAY
Its wrong to always override the state of the machine to running after
reverting to a snapshot. This was only done for the case of user
entering snapshots view from display so lets ensure that indeed is the
case before overriding the state of machine.

With this change, libvirt should ensure that after getting reverted to a
snapshot, machine is back in the state that it was when the snapshot in
question was taken.

https://bugzilla.gnome.org/show_bug.cgi?id=735688