installer-media: prepare_for_installation() throws no error
Since the only real implementation of prepare_for_installation() (in
UnattendedInstaller) handles errors itself anyway and error is ingorable
(and is ignored), there is no need to burdon the caller (VMCreator) with
the error as well. The error catching code in VMCreator was also
assuming internal details of this method so now it doesn't have to.