snapshots-list-row: Always display msg when reverting
We were not displaying "Reverting to SNAPSHOT_NAME" message and spinner
if machine was not in running state. The rationale was that in
non-running state, reverting is a very quick operation and therefore the
stack page switch animation looked weird.

However reverting is not always very fast in this case and the missing
indication is a bit of a UX failure when that happens. Moreover, the
quick page switch doesn't look that weird after all. If we decide to
still address the latter, we can always introduce a very short artificial
delay.