libvirt-machine-props: Ensure name change is saved
The two different UI components (name property and toolbar title) that
can be used to change the name of machine were using two different
methods to do so and hence changing the name from toolbar title wasn't
getting the name change saved.

This patch unifies the method to change the name and ensures that change
of name is saved.

https://bugzilla.gnome.org/show_bug.cgi?id=736096