snapshot-list-row: Remove redundant revealer
There is no need to use a revealer to allow for undoing deletion of
snapshots. Instead we can just remove the snapshot row from the view and
if user clicks undo button, we can simply add it back in.

Cherry-picking this patch from master to 3.14 branch since it turns out
to solve a crash when deleting a snapshot.