display-page: Show display size on widget resize
This is similar to what gnome-terminal does but unlike gnome-terminal we
don't show display size label on user initiating the window drag but
rather *after* every resize of the widget. The label is automatically
hiden away after 4 seconds of last resize operation.

The reason we do it differently is mainly because I failed to find which
signals/events gnome-terminal (vte actually) hooks on to for this.
Besides this isn't so bad solution either and much better than not
showing the label at all.

https://bugzilla.gnome.org/show_bug.cgi?id=738760