spice-display: Remove 'Resize guest' property
Tell SPICE to always try to resize the guest display to allocated size,
instead of exposing this internal detail to users. Not only is this
property non-obvious to most people, we have had some weird side-effects
from people disabling this property in the past.

See the bug below for discussion leading to removal of this option.

https://bugzilla.gnome.org/show_bug.cgi?id=729700