properties: Specify min content size on ScrolledWindow
We are about to move propertiew view into its own window thats made
transient over the main window. If minimum width and height are not set,
ScrolledWindow will hide most of the content once that change is in
place.

So in this patch, we set minimum width to be 640 and minimum height to
be 480. This should be harmless except for screens with extremely low
resolution, in which case I'm sure this won't be our biggest worry
anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=733367