libvirt-machine-props: Use notifications for errors
Instead of launching a dialog on failing to save troubleshooting log, use
notifications.

Known issue: The troubleshooting log is itself a dialog that hides the
properties window at the moment so the notification is hidden unless
user closes the dialog soon (a few seconds) after choosing the file.
This issue will be resolved after we merge troubleshooting log UI into
the properties window itself.

https://bugzilla.gnome.org/show_bug.cgi?id=741047