wizard: Make use of new customization view/API
Instead of launching properties view on user hitting 'Customize' button,
show the new resource customization UI. This means that now user can
only customize resources before hitting 'Create' button but that is fine
because resources are typically the only properties that users want to
customize before machine is launched for the first time.

Also this bebavior/UI is according to the new mockups.

https://bugzilla.gnome.org/show_bug.cgi?id=741046