collection-view: Hide items until wizard is finished
In case of local VMs, we create the machine already before user is out
of wizard. However, we don't want this newly added machine to be visible
until user has hit 'Create' button and/or is out of wizard state.

This patch defers the addition of new items to collection view until we
are out of wizard state.

https://bugzilla.gnome.org/show_bug.cgi?id=741340