Add KeysInputPopover
This popover is meant to allow users to easily send key combos to guest
that would normally be captured by host/client system. The supported
combos are kept limited to not clutter the UI. Currently supported
comboxes are:

* Ctrl+Alt+Backspace
* Ctrl+Alt+F1
* Ctrl+Alt+F2
* Ctrl+Alt+F7

The point of supporting these 3 specific FN combos is to allow user to
easily switch to/from console VT from/to graphics one on X-running
guests. Once they are on a console VT, they can then use Alt+FN combo to
get to other VTs but we don't provide a combo for that since its not
expected to be the most typical use case.

https://bugzilla.gnome.org/show_bug.cgi?id=725303